requirements.txt
azure-common       ==1.1.28
azure-core         ==1.27.1
azure-mgmt-core    ==1.4.0
azure-mgmt-storage ==21.0.0
azure-storage-blob ==12.16.0
backcall           ==0.2.0
certifi            ==2023.5.7
cffi               ==1.15.1
charset-normalizer ==3.1.0
colorama           ==0.4.6
comm               ==0.1.3
contourpy          ==1.1.0
cryptography       ==41.0.1
cycler             ==0.11.0
debugpy            ==1.6.7
decorator          ==5.1.1
executing          ==1.2.0
fonttools          ==4.40.0
greenlet           ==2.0.2
idna               ==3.4
isodate            ==0.6.1
jedi               ==0.18.2
kiwisolver         ==1.4.4
matplotlib         ==3.7.2
matplotlib-inline  ==0.1.6
msrest             ==0.7.1
nest-asyncio       ==1.5.6
numpy              ==1.25.0
oauthlib           ==3.2.2
packaging          ==23.1
pandas             ==2.0.2
parso              ==0.8.3
pickleshare        ==0.7.5
pip                ==23.1.2
platformdirs       ==3.8.0
prompt-toolkit     ==3.0.38
psutil             ==5.9.5
pure-eval          ==0.2.2
pycparser          ==2.21
Pygments           ==2.15.1
pyodbc             ==4.0.39
pyparsing          ==3.0.9
python-dateutil    ==2.8.2
pytz               ==2023.3
pywin32            ==306
pyzmq              ==25.1.0
requests           ==2.31.0
requests-oauthlib  ==1.3.1
scipy              ==1.11.0
setuptools         ==67.8.0
six                ==1.16.0
SQLAlchemy         ==2.0.17
stack-data         ==0.6.2
tornado            ==6.3.2
traitlets          ==5.9.0
typing_extensions  ==4.6.3
tzdata             ==2023.3
urllib3            ==2.0.3
wcwidth            ==0.2.6
wheel              ==0.40.0
